Add Test Data
This commit includes a small (~1.6 MB) BAM file containing reads from
the DMD gene that we can use for testing.  It also includes a file,
test_resources.py which contains code to obtain the location of the
file along with tests that the file exists and that it can be parsed by
the pysam package.