Add Cython Version Requirement
Ari ran into an issue where his cython version with Anaconda could not
correctly compile pysamtools.

Pysamtools specifies its version as cython>=0.24.1, so in theory this
version should work.  However, there may have been a version upgrade
that broke pysamtools, there may be a conflict with anaconda, or there
may be some other issue with the requirement resolution.

We don’t have time to track that down, but as Ari was able to compile
with the latest version, we’ll just set a requirement on that to help
avoid this issue in the future.