Major API Changes
I wanted to both clean up the repo in preparation for release, as well
as explore the code more as we work on the paper.  This pull contains
multiple major changes, including:

1 - Adds code and a command line argument to generate BAM files from a
BED file, so we can do end-to-end checking without actually committing
multiple BAM files files

2 - Removes the dependency on genepeeks-science.  This will allow the
repo to be more broadly accessible, but loses some of some useful
functionality, will have to add this in later

3 - "Classify" how we handle targets, I started using Target objects
inside TargetCollection classes.

4 - Add a full workflow from making the matrix to training the model
and evaluating the sample to the tests.